+++ Firebend.AutoCrud.EntityFramework.Elastic/Extensions/ElasticEntityFrameworkExtensions.cs
@@ -1,6 +1,7 @@
 using System;
 using Firebend.AutoCrud.Core.Interfaces.Models;
 using Firebend.AutoCrud.EntityFramework.Elastic.Implementations;
+using Firebend.AutoCrud.EntityFramework.Elastic.Implementations.Abstractions;
 using Firebend.AutoCrud.EntityFramework.Elastic.Models;
 using Firebend.AutoCrud.EntityFramework.Interfaces;
 
@@ -46,7 +47,7 @@ namespace Firebend.AutoCrud.EntityFramework.Elastic.Extensions
 
             config.Builder.WithRegistration<
                 IEntityFrameworkDbUpdateExceptionHandler<TKey, TEntity>,
-                ConstraintUpdateExceptionHandler<TKey, TEntity>>();
+                AbstractConstraintUpdateExceptionHandler<TKey, TEntity>>();
 
             configure(config);
