+++ Firebend.AutoCrud.EntityFramework/DbContextOptionsProvider.cs
@@ -3,7 +3,7 @@ using Firebend.AutoCrud.Core.Interfaces.Models;
 using Firebend.AutoCrud.EntityFramework.Interfaces;
 using Microsoft.EntityFrameworkCore;
 
-namespace Firebend.AutoCrud.EntityFramework.Abstractions.Client
+namespace Firebend.AutoCrud.EntityFramework
 {
     public class DbContextOptionsProvider<TKey, TEntity> : IDbContextOptionsProvider<TKey, TEntity>
         where TKey : struct
