+++ Firebend.AutoCrud.Tests/Firebend.AutoCrud.Tests.csproj
@@ -11,7 +11,7 @@
     <ItemGroup>
         <PackageReference Include="FluentAssertions" Version="5.10.3" />
         <PackageReference Include="GitHubActionsTestLogger" Version="1.1.2" />
-        <PackageReference Include="nunit" Version="3.13.0" />
+        <PackageReference Include="nunit" Version="3.13.1" />
         <PackageReference Include="NUnit3TestAdapter" Version="3.17.0" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.8.3" />
     </ItemGroup>
