+++ Firebend.AutoCrud.Web/Firebend.AutoCrud.Web.csproj
@@ -22,6 +22,7 @@
       <PackageReference Include="Microsoft.AspNetCore.JsonPatch" Version="5.0.2" />
       <PackageReference Include="Microsoft.AspNetCore.Mvc.Core" Version="2.2.5" />
       <PackageReference Include="Swashbuckle.AspNetCore.Annotations" Version="5.6.3" />
+      <PackageReference Include="System.ComponentModel.Annotations" Version="5.0.0" />
     </ItemGroup>
 
     <ItemGroup>
